+++ src/main/java/de/delautrer/game/blocks/PlantBlock.java
@@ -10,6 +10,7 @@ import de.delautrer.game.loot.LootTable;
 import de.delautrer.game.loot.LootTableManager;
 import de.delautrer.game.items.ItemStack;
 import de.delautrer.game.entity.ItemEntity;
+import org.joml.Vector3d;
 import org.joml.Vector3f;
 import org.joml.Vector3i;
 
@@ -86,10 +87,10 @@ public class PlantBlock extends Block {
 
                 for (ItemStack stack : drops) {
                     // Position zentriert in der Mitte des Blocks
-                    Vector3f dropPos = new Vector3f(
-                            x + 0.5f,
-                            y + 0.5f,
-                            z + 0.5f
+                    Vector3d dropPos = new Vector3d(
+                            x + 0.5,
+                            y + 0.5,
+                            z + 0.5
                     );
 
                     // Ein kleiner Schubs nach oben und zufällig zur Seite ("Plopp"-Effekt)
